requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
psycopg2-binary
